+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+### 0.0.2 (2025 - 08-02)
+
+#### Features
+- Implementation roles with entity, repository and logic
+- Implementation users with entity, repository and logic
+- Initial audits for users
+
 ### 0.0.1 (2025 - 08-02)
 
 #### Features